// server/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Plan = {
  FREE: 'FREE',
  PRO: 'PRO',
  TEAM: 'TEAM'
} as const

export type Plan = (typeof Plan)[keyof typeof Plan]


export const SubscriptionStatus = {
  INACTIVE: 'INACTIVE',
  INCOMPLETE: 'INCOMPLETE',
  INCOMPLETE_EXPIRED: 'INCOMPLETE_EXPIRED',
  TRIALING: 'TRIALING',
  ACTIVE: 'ACTIVE',
  PAST_DUE: 'PAST_DUE',
  UNPAID: 'UNPAID',
  PAUSED: 'PAUSED',
  CANCELLED: 'CANCELLED'
} as const

export type SubscriptionStatus = (typeof SubscriptionStatus)[keyof typeof SubscriptionStatus]


export const ResumeStatus = {
  UPLOADED: 'UPLOADED',
  EXTRACTED: 'EXTRACTED',
  PROCESSING: 'PROCESSING',
  READY: 'READY',
  FAILED: 'FAILED'
} as const

export type ResumeStatus = (typeof ResumeStatus)[keyof typeof ResumeStatus]


export const AnalysisType = {
  RESUME_REVIEW: 'RESUME_REVIEW',
  JOB_MATCH: 'JOB_MATCH',
  COVER_LETTER: 'COVER_LETTER',
  INTERVIEW_PREP: 'INTERVIEW_PREP',
  SALARY_INSIGHT: 'SALARY_INSIGHT'
} as const

export type AnalysisType = (typeof AnalysisType)[keyof typeof AnalysisType]


export const AnalysisStatus = {
  PENDING: 'PENDING',
  PROCESSING: 'PROCESSING',
  COMPLETED: 'COMPLETED',
  FAILED: 'FAILED'
} as const

export type AnalysisStatus = (typeof AnalysisStatus)[keyof typeof AnalysisStatus]


export const ApplicationStatus = {
  SAVED: 'SAVED',
  APPLIED: 'APPLIED',
  SCREENING: 'SCREENING',
  INTERVIEW: 'INTERVIEW',
  TECHNICAL_TEST: 'TECHNICAL_TEST',
  FINAL_INTERVIEW: 'FINAL_INTERVIEW',
  OFFER: 'OFFER',
  REJECTED: 'REJECTED',
  WITHDRAWN: 'WITHDRAWN'
} as const

export type ApplicationStatus = (typeof ApplicationStatus)[keyof typeof ApplicationStatus]


export const InterviewDifficulty = {
  BEGINNER: 'BEGINNER',
  INTERMEDIATE: 'INTERMEDIATE',
  ADVANCED: 'ADVANCED'
} as const

export type InterviewDifficulty = (typeof InterviewDifficulty)[keyof typeof InterviewDifficulty]


export const UsageType = {
  RESUME_ANALYSIS: 'RESUME_ANALYSIS',
  JOB_COMPARISON: 'JOB_COMPARISON',
  INTERVIEW_PREP: 'INTERVIEW_PREP',
  COVER_LETTER: 'COVER_LETTER'
} as const

export type UsageType = (typeof UsageType)[keyof typeof UsageType]